Updated library to support Visual Recognition V2
