--- src/main/java/com/ibm/watson/developer_cloud/visual_recognition/v1/model/Label.java
@@ -1,69 +1,0 @@
-/**
- * Copyright 2015 IBM Corp. All Rights Reserved.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * 
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package com.ibm.watson.developer_cloud.visual_recognition.v1.model;
-
-import com.google.gson.annotations.SerializedName;
-import com.ibm.watson.developer_cloud.service.model.GenericModel;
-
-/**
- * The Class Label.
- */
-public class Label extends GenericModel {
-
-  /** The name. */
-  @SerializedName("label_name")
-  private String name;
-
-  /** The score. */
-  @SerializedName("label_score")
-  private double score;
-
-  /**
-   * Gets the name.
-   * 
-   * @return the name
-   */
-  public String getName() {
-    return name;
-  }
-
-  /**
-   * Gets the score.
-   * 
-   * @return the score
-   */
-  public double getScore() {
-    return score;
-  }
-
-  /**
-   * Sets the name.
-   * 
-   * @param name the new name
-   */
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  /**
-   * Sets the score.
-   * 
-   * @param score the new score
-   */
-  public void setScore(double score) {
-    this.score = score;
-  }
-
-}
