+++ src/main/java/com/ibm/watson/developer_cloud/visual_recognition/v2/model/Classifier.java
@@ -1,0 +1,73 @@
+/**
+ * Copyright 2015 IBM Corp. All Rights Reserved.
+ * <p/>
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ * <p/>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p/>
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package com.ibm.watson.developer_cloud.visual_recognition.v2.model;
+
+import com.google.gson.annotations.SerializedName;
+import com.ibm.watson.developer_cloud.service.model.GenericModel;
+
+/**
+ * The Class Classifier.
+ */
+public class Classifier extends GenericModel {
+
+    /**
+     * The name.
+     */
+    @SerializedName("classifier_name")
+    private String name;
+
+    /**
+     * The score.
+     */
+    @SerializedName("classifier_score")
+    private double score;
+
+    /**
+     * Gets the name.
+     *
+     * @return the name
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * Gets the score.
+     *
+     * @return the score
+     */
+    public double getScore() {
+        return score;
+    }
+
+    /**
+     * Sets the name.
+     *
+     * @param name the new name
+     */
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    /**
+     * Sets the score.
+     *
+     * @param score the new score
+     */
+    public void setScore(double score) {
+        this.score = score;
+    }
+
+}
