+++ src/main/java/com/ibm/watson/developer_cloud/visual_recognition/v2/model/package-info.java
@@ -11,5 +11,5 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.ibm.watson.developer_cloud.visual_recognition.v1;
+package com.ibm.watson.developer_cloud.visual_recognition.v2.model;
